pyproject.toml
[project]
name = "compreensive-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "haystack-ai>=2.25.2",
    "judge0>=0.0.8",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
